refactor: rename namespace stacked to doc vec
